[FrontController] merged Controller and FrontController components for more coherence #388
