+++ Rest/Tests/EventListener/ValidationListenerTest.php
@@ -25,7 +25,7 @@ namespace BackBee\Rest\Tests\EventListener;
 
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpKernel\Event\FilterControllerEvent;
-use BackBee\FrontController\FrontController;
+use BackBee\Controller\FrontController;
 use BackBee\Rest\EventListener\ValidationListener;
 use BackBee\Rest\Tests\Fixtures\Controller\FixtureAnnotatedController;
 use BackBee\Tests\TestCase;
